+++ tests/integration/test_fim/test_registry/test_registry_report_changes/test_disk_quota/test_registry_disk_quota_values.py
@@ -1,4 +1,4 @@
-'''
+"""
 copyright: Copyright (C) 2015-2021, Wazuh Inc.
 
            Created by Wazuh, Inc. <info@wazuh.com>.
@@ -53,16 +53,18 @@ pytest_args:
 
 tags:
     - fim_registry_report_changes
-'''
+"""
 import os
 
 import pytest
-from test_fim.test_files.test_report_changes.common import generate_string
 from wazuh_testing import global_parameters
-from wazuh_testing.fim import LOG_FILE_PATH, registry_value_cud, KEY_WOW64_32KEY, KEY_WOW64_64KEY, generate_params, \
-    calculate_registry_diff_paths, create_registry, delete_registry, registry_parser, \
-    check_time_travel
-from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
+from wazuh_testing.fim import (LOG_FILE_PATH, KEY_WOW64_32KEY, KEY_WOW64_64KEY, generate_params,
+                               calculate_registry_diff_paths, registry_value_create, registry_value_update,
+                               registry_value_delete, registry_parser, create_values_content)
+from wazuh_testing.fim_module import (WINDOWS_HKEY_LOCAL_MACHINE, MONITORED_KEY, MONITORED_KEY_2,
+                                                    SIZE_LIMIT_CONFIGURED_VALUE, ERR_MSG_CONTENT_CHANGES_EMPTY,
+                                                    ERR_MSG_CONTENT_CHANGES_NOT_EMPTY)
+from wazuh_testing.tools.configuration import load_wazuh_configurations
 from wazuh_testing.tools.monitoring import FileMonitor
 
 # Marks
@@ -71,60 +73,57 @@ pytestmark = [pytest.mark.win32, pytest.mark.tier(level=1)]
 
 # Variables
 
-key = "HKEY_LOCAL_MACHINE"
-sub_key_1 = "SOFTWARE\\key1"
-sub_key_2 = "SOFTWARE\\key2"
-
-test_regs = [os.path.join(key, sub_key_1),
-             os.path.join(key, sub_key_2)]
-
-test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), 'data')
+test_regs = [os.path.join(WINDOWS_HKEY_LOCAL_MACHINE, MONITORED_KEY), 
+             os.path.join(WINDOWS_HKEY_LOCAL_MACHINE, MONITORED_KEY_2)]
+test_data_path = os.path.join(os.path.dirname(os.path.realpath(__file__)), "data")
 wazuh_log_monitor = FileMonitor(LOG_FILE_PATH)
-reg1, reg2 = test_regs
-size_limit_configured = 10 * 1024
+scan_delay = 2
 
 # Configurations
 
-p, m = generate_params(modes=['scheduled'], extra_params={'WINDOWS_REGISTRY_1': reg1,
-                                                          'WINDOWS_REGISTRY_2': reg2,
-                                                          'FILE_SIZE_ENABLED': 'no',
-                                                          'FILE_SIZE_LIMIT': '10KB',
-                                                          'DISK_QUOTA_ENABLED': 'yes',
-                                                          'DISK_QUOTA_LIMIT': '1KB'})
+params, metadata = generate_params(modes=["scheduled"], extra_params={
+                                                        "WINDOWS_REGISTRY_1": test_regs[0],
+                                                        "WINDOWS_REGISTRY_2": test_regs[1],
+                                                        "FILE_SIZE_ENABLED": "no",
+                                                        "FILE_SIZE_LIMIT": "10KB",
+                                                        "DISK_QUOTA_ENABLED": "yes",
+                                                        "DISK_QUOTA_LIMIT": "1KB",
+                                                        })
 
-configurations_path = os.path.join(test_data_path, 'wazuh_registry_report_changes_limits_quota.yaml')
-
-configurations = load_wazuh_configurations(configurations_path, __name__, params=p, metadata=m)
+configurations_path = os.path.join(test_data_path, "wazuh_registry_report_changes_limits_quota.yaml")
+configurations = load_wazuh_configurations(configurations_path, __name__, params=params, metadata=metadata)
 
 
 # Fixtures
 
 
-@pytest.fixture(scope='module', params=configurations)
+@pytest.fixture(scope="module", params=configurations)
 def get_configuration(request):
     """Get configurations from the module."""
     return request.param
 
 
-@pytest.mark.parametrize('size', [
-    (4 * 1024),
-    (32 * 1024)
-])
-@pytest.mark.parametrize('key, subkey, arch, value_name, tags_to_apply', [
-    (key, sub_key_1, KEY_WOW64_64KEY, "some_value", {'test_limits'}),
-    (key, sub_key_1, KEY_WOW64_32KEY, "some_value", {'test_limits'}),
-    (key, sub_key_2, KEY_WOW64_64KEY, "some_value", {'test_limits'})
-])
-def test_disk_quota_values(key, subkey, arch, value_name, tags_to_apply, size,
-                           get_configuration, configure_environment, restart_syscheckd,
-                           wait_for_fim_start):
-    '''
+@pytest.mark.parametrize("size", [(4096), (32768)])
+@pytest.mark.parametrize("key, subkey, arch, value_name",
+    [
+        (WINDOWS_HKEY_LOCAL_MACHINE, MONITORED_KEY, KEY_WOW64_64KEY, "some_value"),
+        (WINDOWS_HKEY_LOCAL_MACHINE, MONITORED_KEY, KEY_WOW64_32KEY, "some_value"),
+        (WINDOWS_HKEY_LOCAL_MACHINE, MONITORED_KEY_2, KEY_WOW64_64KEY, "some_value"),
+    ],
+)
+def test_disk_quota_values(key, subkey, arch, value_name, size, get_configuration, configure_environment,
+                           restart_syscheckd, wait_for_fim_start):
+    """
     description: Check if the 'wazuh-syscheckd' daemon limits the size of the folder where the data used
                  to perform the 'diff' operations is stored when the 'disk_quota' limit is set. For this
                  purpose, the test will monitor a key, create a testing value smaller than the 'disk_quota'
                  limit, and increase its size on each test case. Finally, the test will verify that the
                  compressed file has been created, and the related FIM event includes the 'content_changes'
-                 field if the value size does not exceed the specified limit and vice versa.
+                 field if the value size does not exceed the specified limit and viceversa.
+                 - Case 1: small file - when compressed it will be less than the disk_quota. The file is generated 
+                 and the logs have content_changes data.
+                 - Case 2: big size - when compressed the file would be bigger than the disk_quota. The file is not
+                 generated and the logs should not have content_changes data.
 
     wazuh_min_version: 4.2.0
 
@@ -141,9 +140,6 @@ def test_disk_quota_values(key, subkey, arch, value_name, tags_to_apply, size,
         - value_name:
             type: str
             brief: Name of the testing value that will be created
-        - tags_to_apply:
-            type: set
-            brief: Run test if matches with a configuration identifier, skip otherwise.
         - size:
             type: int
             brief: Size of the content to write in the testing value.
@@ -155,7 +151,7 @@ def test_disk_quota_values(key, subkey, arch, value_name, tags_to_apply, size,
             brief: Configure a custom environment for testing.
         - restart_syscheckd:
             type: fixture
-            brief: Clear the 'ossec.log' file and start a new monitor.
+            brief: Clear the Wazuh logs file and start a new monitor.
         - wait_for_fim_start:
             type: fixture
             brief: Wait for realtime start, whodata start, or end of initial FIM scan.
@@ -176,34 +172,32 @@ def test_disk_quota_values(key, subkey, arch, value_name, tags_to_apply, size,
 
     tags:
         - scheduled
-        - time_travel
-    '''
-    check_apply_test(tags_to_apply, get_configuration['tags'])
-    value_content = generate_string(size, '0')
-    values = {value_name: value_content}
+    """
+    values = create_values_content(value_name, size)
 
     _, diff_file = calculate_registry_diff_paths(key, subkey, arch, value_name)
 
     def report_changes_validator_no_diff(event):
-        """Validate content_changes attribute exists in the event"""
-        assert event['data'].get('content_changes') is None, 'content_changes isn\'t empty'
+        """Validate content_changes attribute does not exist in the event"""
+        assert event["data"].get("content_changes") is None, ERR_MSG_CONTENT_CHANGES_NOT_EMPTY
 
     def report_changes_validator_diff(event):
         """Validate content_changes attribute exists in the event"""
-        assert os.path.exists(diff_file), '{diff_file} does not exist'
-        assert event['data'].get('content_changes') is not None, 'content_changes is empty'
+        assert os.path.exists(diff_file), "{diff_file} does not exist"
+        assert event["data"].get("content_changes") is not None, ERR_MSG_CONTENT_CHANGES_EMPTY
 
-    if size > size_limit_configured:
+    if size > SIZE_LIMIT_CONFIGURED_VALUE:
         callback_test = report_changes_validator_no_diff
     else:
         callback_test = report_changes_validator_diff
 
-    create_registry(registry_parser[key], subkey, arch)
-
-    registry_value_cud(key, subkey, wazuh_log_monitor, arch=arch, value_list=values,
-                       time_travel=get_configuration['metadata']['fim_mode'] == 'scheduled',
-                       min_timeout=global_parameters.default_timeout, triggers_event=True,
-                       validators_after_update=[callback_test])
-
-    delete_registry(registry_parser[key], subkey, arch)
-    check_time_travel(True, monitor=wazuh_log_monitor)
+    # Create the value inside the key - we do it here because it key or arch is not known before the test launches
+    registry_value_create(key, subkey, wazuh_log_monitor, arch=arch, value_list=values, wait_for_scan=True,
+                          scan_delay=scan_delay, min_timeout=global_parameters.default_timeout, triggers_event=True)
+    # Modify the value to check if the diff file is generated or not, as expected
+    registry_value_update(key, subkey, wazuh_log_monitor, arch=arch, value_list=values, wait_for_scan=True,
+                          scan_delay=scan_delay, min_timeout=global_parameters.default_timeout,   triggers_event=True,
+                          validators_after_update=[callback_test])
+    # Delete the vaue created to clean up enviroment
+    registry_value_delete(key, subkey, wazuh_log_monitor, arch=arch, value_list=values, wait_for_scan=True,
+                          scan_delay=scan_delay, min_timeout=global_parameters.default_timeout, triggers_event=True)