+++ tests/integration/test_fim/test_registry/test_registry_report_changes/test_registry_file_size_default.py
@@ -58,13 +58,13 @@ import os
 
 import pytest
 from wazuh_testing import global_parameters
-from wazuh_testing.fim import LOG_FILE_PATH, KEY_WOW64_32KEY, KEY_WOW64_64KEY, generate_params, \
-    callback_diff_size_limit_value, create_registry, registry_parser, modify_registry_value, \
+from wazuh_testing.fim import LOG_FILE_PATH, KEY_WOW64_32KEY, KEY_WOW64_64KEY, generate_params, create_registry, registry_parser, modify_registry_value, \
     check_time_travel, validate_registry_value_event, callback_value_event, REG_SZ
 from wazuh_testing.tools.configuration import load_wazuh_configurations, check_apply_test
 from wazuh_testing.tools.file import truncate_file
-from wazuh_testing.tools.monitoring import FileMonitor
+from wazuh_testing.tools.monitoring import FileMonitor, generate_monitoring_callback
 from wazuh_testing.tools.services import control_service
+from wazuh_testing.fim_module.fim_variables import CB_MAXIMUM_FILE_SIZE
 
 # Marks
 
@@ -116,7 +116,7 @@ def get_configuration(request):
     (key, sub_key_1, KEY_WOW64_32KEY, "some_value", {'test_report_changes'}),
     (key, sub_key_2, KEY_WOW64_64KEY, "some_value", {'test_report_changes'})
 ])
-@pytest.mark.skip(reason="It will be blocked by #1602, when it was solve we can enable again this test")
+@pytest.mark.skip(reason="It will be blocked by #2174, when it was solve we can enable again this test")
 def test_file_size_default(key, subkey, arch, value_name, tags_to_apply,
                            get_configuration, configure_environment, restart_syscheckd_each_time):
     '''
@@ -178,7 +178,7 @@ def test_file_size_default(key, subkey, arch, value_name, tags_to_apply,
     mode = get_configuration['metadata']['fim_mode']
 
     file_size_values = wazuh_log_monitor.start(timeout=global_parameters.default_timeout,
-                                               callback=callback_diff_size_limit_value,
+                                               callback=generate_monitoring_callback(CB_MAXIMUM_FILE_SIZE),
                                                accum_results=3,
                                                error_message='Did not receive expected '
                                                              '"Maximum file size limit to generate diff information '